requirements.txt
anthropic>=0.18.0
tencentcloud-sdk-python>=3.0.1090
requests>=2.31.0
